Move polar_ft.py to operators dir.
