+++ src/aspire/basis/__init__.py
@@ -15,6 +15,5 @@ from .fle_2d import FLEBasis2D
 from .fpswf_2d import FPSWFBasis2D
 from .fpswf_3d import FPSWFBasis3D
 from .fspca import FSPCABasis
-from .polar_2d import PolarFT
 from .pswf_2d import PSWFBasis2D
 from .pswf_3d import PSWFBasis3D
