+++ src/aspire/operators/__init__.py
@@ -18,4 +18,5 @@ from .filters import (
     ZeroFilter,
     evaluate_src_filters_on_grid,
 )
+from .polar_ft import PolarFT
 from .wemd import wemd_embed, wemd_norm
