+++ tests/test_polar_ft.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from aspire.basis import PolarFT
+from aspire.operators import PolarFT
 from aspire.image import Image
 from aspire.utils import gaussian_2d, grid_2d
 from aspire.volume import AsymmetricVolume, CnSymmetricVolume
